Possible compromise?
Possible compromise for now? I still like the idea of a graphical representation as that works best for me at a quick glance, but I definitely agree with the issues noted above in regards to readability of the impact glyphs in different fontsets, as well as the possibility of issues with regards to translation across different langauges, which a text legend absolutely aids. I think I might need to create custom "roles" to colorize the graph, and/or investigate other glyphs entirely which may hopefully render much nicer.