Update Singularity Documentation
The command for running Singularity was incorrect. Additionally, I removed excess Options and Flags which are unnecesary for running this command. This makes it similar in structure to the command for running docker listed here: https://www.opendronemap.org/odm/

Removed the command for running Singularity with ODM, as it was incorrect and was a duplicate of the other command. 

Fixed link to singularity website.